added unit test for page change
